refactor network pacakge
