+++ network/p2p/p2p_pubsub.go
@@ -1,0 +1,161 @@
+package p2p
+
+import (
+	"fmt"
+	"github.com/herumi/bls-eth-go-binary/bls"
+	pubsub "github.com/libp2p/go-libp2p-pubsub"
+	"github.com/pkg/errors"
+	"go.uber.org/zap"
+	"strings"
+)
+
+// SubscribeToValidatorNetwork  for new validator create new topic, subscribe and start listen
+func (n *p2pNetwork) SubscribeToValidatorNetwork(validatorPk *bls.PublicKey) error {
+	n.psTopicsLock.Lock()
+	defer n.psTopicsLock.Unlock()
+
+	pubKey := validatorPk.SerializeToHexStr()
+
+	if _, ok := n.cfg.Topics[pubKey]; !ok {
+		if err := n.joinTopic(pubKey); err != nil {
+			return errors.Wrap(err, "failed to join to topic")
+		}
+	}
+
+	if !n.psSubscribedTopics[pubKey] {
+		sub, err := n.cfg.Topics[pubKey].Subscribe()
+		if err != nil {
+			if err != pubsub.ErrTopicClosed {
+				return errors.Wrap(err, "failed to subscribe on Topic")
+			}
+			// rejoin a topic in case it was closed, and trying to subscribe again
+			if err := n.joinTopic(pubKey); err != nil {
+				return errors.Wrap(err, "failed to join to topic")
+			}
+			sub, err = n.cfg.Topics[pubKey].Subscribe()
+			if err != nil {
+				return errors.Wrap(err, "failed to subscribe on Topic")
+			}
+		}
+		n.psSubscribedTopics[pubKey] = true
+		go func() {
+			n.listen(sub)
+			// mark topic as not subscribed
+			n.psTopicsLock.Lock()
+			defer n.psTopicsLock.Unlock()
+			n.psSubscribedTopics[pubKey] = false
+		}()
+	}
+
+	return nil
+}
+
+// AllPeers returns all connected peers for a validator PK (except for the validator itself)
+func (n *p2pNetwork) AllPeers(validatorPk []byte) ([]string, error) {
+	topic, err := n.getTopic(validatorPk)
+	if err != nil {
+		return nil, err
+	}
+
+	return n.allPeersOfTopic(topic), nil
+}
+
+// joinTopic joins to the given topic and mark it in topics map
+// this method is not thread-safe - should be called after psTopicsLock was acquired
+func (n *p2pNetwork) joinTopic(pubKey string) error {
+	topic, err := n.pubsub.Join(getTopicName(pubKey))
+	if err != nil {
+		return errors.Wrap(err, "failed to join to topic")
+	}
+	n.cfg.Topics[pubKey] = topic
+	return nil
+}
+
+// closeTopic closes the given topic
+func (n *p2pNetwork) closeTopic(topicName string) error {
+	n.psTopicsLock.RLock()
+	defer n.psTopicsLock.RUnlock()
+
+	pk := unwrapTopicName(topicName)
+	if t, ok := n.cfg.Topics[pk]; ok {
+		delete(n.cfg.Topics, pk)
+		return t.Close()
+	}
+	return nil
+}
+
+// getTopic return topic by validator public key
+func (n *p2pNetwork) getTopic(validatorPK []byte) (*pubsub.Topic, error) {
+	n.psTopicsLock.RLock()
+	defer n.psTopicsLock.RUnlock()
+
+	if validatorPK == nil {
+		return nil, errors.New("ValidatorPk is nil")
+	}
+	topic := n.fork.ValidatorTopicID(validatorPK)
+	if _, ok := n.cfg.Topics[topic]; !ok {
+		return nil, errors.New("topic is not exist or registered")
+	}
+	return n.cfg.Topics[topic], nil
+}
+
+// AllPeers returns all connected peers for a validator PK (except for the validator itself and public peers like exporter)
+func (n *p2pNetwork) allPeersOfTopic(topic *pubsub.Topic) []string {
+	ret := make([]string, 0)
+
+	skippedPeers := map[string]bool{
+		n.cfg.ExporterPeerID: true,
+	}
+
+	for _, p := range topic.ListPeers() {
+		if s := peerToString(p); !skippedPeers[s] {
+			ret = append(ret, peerToString(p))
+		}
+	}
+
+	return ret
+}
+
+// listen listens on the given subscription
+func (n *p2pNetwork) listen(sub *pubsub.Subscription) {
+	t := sub.Topic()
+	defer sub.Cancel()
+	n.logger.Info("start listen to topic", zap.String("topic", t))
+	for {
+		select {
+		case <-n.ctx.Done():
+			if err := n.closeTopic(t); err != nil {
+				n.logger.Error("failed to close topic", zap.String("topic", t), zap.Error(err))
+			}
+			n.logger.Info("closed topic", zap.String("topic", t))
+		default:
+			msg, err := sub.Next(n.ctx)
+			if err != nil {
+				n.logger.Error("failed to get message from subscription Topics", zap.Error(err))
+				return
+			}
+			// For debugging
+			n.logger.Debug("received raw network msg", zap.ByteString("network.Message bytes", msg.Data))
+
+			cm, err := n.fork.DecodeNetworkMsg(msg.Data)
+			if err != nil {
+				n.logger.Error("failed to un-marshal message", zap.Error(err))
+				continue
+			}
+			if n.reportLastMsg && len(msg.ReceivedFrom) > 0 {
+				reportLastMsg(msg.ReceivedFrom.String())
+			}
+			n.propagateSignedMsg(cm)
+		}
+	}
+}
+
+// getTopicName return formatted topic name
+func getTopicName(pk string) string {
+	return fmt.Sprintf("%s.%s", topicPrefix, pk)
+}
+
+// getTopicName return formatted topic name
+func unwrapTopicName(topicName string) string {
+	return strings.Replace(topicName, fmt.Sprintf("%s.", topicPrefix), "", 1)
+}
