+++ network/p2p/p2p_sync_test.go
@@ -41,7 +41,7 @@ func TestSyncMessageBroadcastingTimeout(t *testing.T) {
 
 	// create 2 peers
 	peer1, err := New(context.Background(), logger, &Config{
-		DiscoveryType:    "mdns",
+		DiscoveryType:    discoveryTypeMdns,
 		Enr:              "enr:-LK4QMIAfHA47rJnVBaGeoHwXOrXcCNvUaxFiDEE2VPCxQ40cu_k2hZsGP6sX9xIQgiVnI72uxBBN7pOQCo5d9izhkcBh2F0dG5ldHOIAAAAAAAAAACEZXRoMpD1pf1CAAAAAP__________gmlkgnY0gmlwhH8AAAGJc2VjcDI1NmsxoQJu41tZ3K8fb60in7AarjEP_i2zv35My_XW_D_t6Y1fJ4N0Y3CCE4iDdWRwgg-g",
 		UDPPort:          12000,
 		TCPPort:          13000,
@@ -52,7 +52,7 @@ func TestSyncMessageBroadcastingTimeout(t *testing.T) {
 	require.NoError(t, err)
 
 	peer2, err := New(context.Background(), logger, &Config{
-		DiscoveryType:    "mdns",
+		DiscoveryType:    discoveryTypeMdns,
 		Enr:              "enr:-LK4QMIAfHA47rJnVBaGeoHwXOrXcCNvUaxFiDEE2VPCxQ40cu_k2hZsGP6sX9xIQgiVnI72uxBBN7pOQCo5d9izhkcBh2F0dG5ldHOIAAAAAAAAAACEZXRoMpD1pf1CAAAAAP__________gmlkgnY0gmlwhH8AAAGJc2VjcDI1NmsxoQJu41tZ3K8fb60in7AarjEP_i2zv35My_XW_D_t6Y1fJ4N0Y3CCE4iDdWRwgg-g",
 		UDPPort:          12001,
 		TCPPort:          13001,
@@ -81,7 +81,7 @@ func TestSyncMessageBroadcasting(t *testing.T) {
 
 	// create 2 peers
 	peer1, err := New(context.Background(), logger, &Config{
-		DiscoveryType:    "mdns",
+		DiscoveryType:    discoveryTypeMdns,
 		Enr:              "enr:-LK4QMIAfHA47rJnVBaGeoHwXOrXcCNvUaxFiDEE2VPCxQ40cu_k2hZsGP6sX9xIQgiVnI72uxBBN7pOQCo5d9izhkcBh2F0dG5ldHOIAAAAAAAAAACEZXRoMpD1pf1CAAAAAP__________gmlkgnY0gmlwhH8AAAGJc2VjcDI1NmsxoQJu41tZ3K8fb60in7AarjEP_i2zv35My_XW_D_t6Y1fJ4N0Y3CCE4iDdWRwgg-g",
 		UDPPort:          12000,
 		TCPPort:          13000,
@@ -92,7 +92,7 @@ func TestSyncMessageBroadcasting(t *testing.T) {
 	require.NoError(t, err)
 
 	peer2, err := New(context.Background(), logger, &Config{
-		DiscoveryType:    "mdns",
+		DiscoveryType:    discoveryTypeMdns,
 		Enr:              "enr:-LK4QMIAfHA47rJnVBaGeoHwXOrXcCNvUaxFiDEE2VPCxQ40cu_k2hZsGP6sX9xIQgiVnI72uxBBN7pOQCo5d9izhkcBh2F0dG5ldHOIAAAAAAAAAACEZXRoMpD1pf1CAAAAAP__________gmlkgnY0gmlwhH8AAAGJc2VjcDI1NmsxoQJu41tZ3K8fb60in7AarjEP_i2zv35My_XW_D_t6Y1fJ4N0Y3CCE4iDdWRwgg-g",
 		UDPPort:          12001,
 		TCPPort:          13001,
